# server/src/main/java/com/storefront/dto/OrderRequestDTO.java
package com.storefront.dto;

import java.util.List;

public class OrderRequestDTO {
    private Long storeId;
    private List<OrderItemRequestDTO> items;

    public Long getStoreId() {
        return storeId;
    }

    public void setStoreId(Long storeId) {
        this.storeId = storeId;
    }

    public List<OrderItemRequestDTO> getItems() {
        return items;
    }

    public void setItems(List<OrderItemRequestDTO> items) {
        this.items = items;
    }
}